==> 1_training data generation/2-2_Concat_simulation_data.ipynb <==
# Concatenating simulation data

**Author:** Manuela Linke, HTWG Konstanz 

**Date:** 15.03.2024 

**Summary:** This Script concatenates the simulated training data of different simulation runs (from the same time_series_data.feather-file). Necessary for me beauce I used three different Computers to generate the Dataset. 

**License:** Copyright 2024 Manuela Linke

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.

import numpy as np
import os
import datetime as dt
import time as t

input_path = os.path.join('data', 'preprocessed', '')
date1 = '2024-03-22_13-06-27' # Data [0 : 18000]
date2 = '2024-03-23_09-52-55' # Data [18000 : 93000]
date3 = '2024-03-23_12-01-46' # Data [93000 : 120000]

ts = t.time()
newdate = dt.datetime.fromtimestamp(ts).strftime('%Y-%m-%d_%H-%M-%S')

#Load # 1st set of generated data
X_load1 = np.load(input_path + date1 + "_Load_distribution.npy")
X_volt1 = np.load(input_path + date1 + "_voltage_distribution.npy")
y_changes1 = np.load(input_path + date1 + "__min_chg_from_base.npy")

#Load # 2nd set of generated data
X_load2 = np.load(input_path + date2 + "_Load_distribution.npy")
X_volt2 = np.load(input_path + date2 + "_voltage_distribution.npy")
y_changes2 = np.load(input_path + date2 + "__min_chg_from_base.npy")

#Load # 3rd set of generated data
X_load3 = np.load(input_path + date3 + "_Load_distribution.npy")
X_volt3 = np.load(input_path + date3 + "_voltage_distribution.npy")
y_changes3 = np.load(input_path + date3 + "__min_chg_from_base.npy")

# Checking the data
print(X_load1.shape)
print(X_load2.shape)
print(X_load3.shape)

# Concatenate the arrays
X_load_combined = np.concatenate((X_load1, X_load2, X_load3), axis=0)
X_volt_combined = np.concatenate((X_volt1, X_volt2, X_volt3), axis=0)
y_changes_combined = np.concatenate((y_changes1, y_changes2, y_changes3), axis=0)

#Checking the new data
print(X_load_combined.shape)

# saving the new data
folder = os.path.join('data', 'preprocessed')
np.save(folder + "/" + newdate + "_" + "voltage_distribution.npy", X_volt_combined)
np.save(folder + "/" + newdate + "_" + "Load_distribution.npy", X_load_combined)
np.save(folder + "/" + newdate + "_" + "_min_chg_from_base.npy", y_changes_combined)
